profile generation, cleaned up slideshow bug, shantanav
