+++ client/src/components/body/landingpage/index.js
@@ -5,36 +5,15 @@ import handleRedirect from "../../../helpers/Redirect";
 import {actions, actions as mainActions} from "../../../reducers/MainReducer";
 import {bindActionCreators} from "redux";
 import {connect} from "react-redux";
-import Profile from "./citation/Profile";
-import studentInformation from "./citation/studentInfomation";
-import professorInformation from "./citation/professorInformation";
 import labInformation from "./lab/labInformation";
 import Lab from "./lab/lab";
+import ProfileGeneration from "./citation/ProfileGeneration";
 
 const mapDispatchToProps = (dispatch) => {
   return {
     actions: bindActionCreators(mainActions, dispatch)
   };
 };
-
-function renderProfileData(profileInformation) {
-  return profileInformation.map((profileInfo, index) => {
-      const {profile_image, name,title,bio,socials,network} = profileInfo //destructuring
-      return (
-              
-              <Profile 
-                  key={index}
-                  profile_image= {profile_image} 
-                  name= {name} 
-                  title= {title} 
-                  bio={bio}
-                  socials={socials}
-                  network={network}
-              />
-      );
-  })
-}
-
 
 function renderLabData() {
   return labInformation.map((labInfo, index) => {
@@ -53,55 +32,9 @@ function renderLabData() {
   })
 }
 
-function renderSlideset(){
-  let profiles = renderProfileData(studentInformation);
-  let rows =[];
-  let profile_row=[];
-  for(let i in profiles){
-    profile_row.push(profiles[i]);
-    if(profile_row.length ===3){
-      rows.push(profile_row);
-      profile_row=[];
-    } 
-  }
-  if(profile_row.length!==0){
-    rows.push(profile_row);
-  }
-  return rows.map((rows,index)=>{
-    return(
-      <div key={index} alt="students" class="slide active">
-          <div key={index} alt="students" class="landingpage__row">
-            {rows}
-          </div>
-      </div>
-    )})
-}
-
-let slideshowInterval='';
-
-function initSlideShow(slideshow) {
-    const slides = document.querySelectorAll(`#${slideshow.id} [role="list"] .slide`);
-    let index = 0, time = 5000;
-    if(slideshowInterval===''){
-      slides.forEach((slide)=>{
-        slide.classList.remove('active');
-      })
-      slides[index].classList.add('active');
-      slideshowInterval=setInterval( () => {
-        slides[index].classList.remove('active');
-        index++;
-        if (index === slides.length) index = 0;
-        slides[index].classList.add('active');
-    }, time);
-  }
-} 
-
-
 const Home = (props) => {
   
   const {actions} = props;
-  let slideshows = document.querySelectorAll('[data-component="slideshow"]');
-  slideshows.forEach(initSlideShow);
   return (
     <div class="landingpage">
       {/* Header */}
@@ -171,31 +104,8 @@ const Home = (props) => {
       </section>
       {/* Team Citation */}
       <div id="citation"/>
-       <hr class="horiz" />
-      <section class="page-section landingpage__pagesection" >
-        <div class="container" >
-            <div class="row">
-              <div class="col-lg-12 text-center">
-                <h2 class="section-heading text-uppercase">Team Members</h2>
-  
-                <h3 class="section-subheading " >
-                  Meet our team.
-                </h3>
-              </div>
-            </div>
-            <div class="landingpage__row">
-              <div alt="professors" class="landingpage__row">
-                {renderProfileData(professorInformation)}
-              </div>
-            </div>
-  
-            <div id="slideshow" alt="students" class="landingpage__row" data-component="slideshow">
-                <div role="list">
-                  {renderSlideset()}
-                </div>
-            </div>
-        </div>
-      </section>
+      <hr class="horiz" />
+        <ProfileGeneration />
       <hr class="horiz" />
       {/* Clients */}
       <section class="py-5">
